build: Split main, add convert bits
