+++ scripts/source/main_schema.py
@@ -1,0 +1,31 @@
+#!/usr/bin/python3
+#
+# Copyright (c) 2026-2026 Google, Inc.
+# Copyright (C) 2026-2026 Valve Corporation
+# Copyright (c) 2026-2026 LunarG, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License")
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Authors: 
+# - Christophe Riccio <christophe@lunarg.com>
+
+from source.vulkan_object_utils import initVulkanObject
+from source.generate_profiles_schema import VulkanProfilesSchemaGenerator2
+
+def main_schema(args):
+    if args.registry is None:
+        vk = initVulkanObject(args.api)
+    else:
+        vk = initVulkanObject(args.api, args.registry, True)
+    generator = VulkanProfilesSchemaGenerator2(vk)
+    generator.generate(args.output)
